requirements.txt
python-dotenv
requests==2.32.5